+++ src/main/java/com/sandymandy/pleasurecraft/util/SceneOptions.java
@@ -10,31 +10,31 @@ import java.util.List;
 
 public record SceneOptions(
         String name,
+        int requiredRelationshipLevel,
         List<String> introAnim,
         List<String> slowAnim,
         List<String> fastAnim,
         String cumAnim,
+        boolean needsToStrip,
         boolean isBedScene,
         float bedOffset,
-        List<String> bedIdle,
-        boolean needsToStrip,
-        int requiredRelationshipLevel
+        List<String> bedIdle
 
 ) {
 
-    public static final SceneOptions EMPTY = new SceneOptions("",new ArrayList<>(),new ArrayList<>(),new ArrayList<>(),"",false,0f, new ArrayList<>(), false, 0);
+    public static final SceneOptions EMPTY = new SceneOptions("", 0,new ArrayList<>(),new ArrayList<>(),new ArrayList<>(),"",false, false, 0f, new ArrayList<>());
 
     public static final PacketCodec<RegistryByteBuf, SceneOptions> CODEC = PacketCodecExtra.tuple(
             PacketCodecs.STRING, SceneOptions::name,
+            PacketCodecs.VAR_INT, SceneOptions::requiredRelationshipLevel,
             PacketCodecs.collection(ArrayList::new , PacketCodecs.STRING), SceneOptions::introAnim,
             PacketCodecs.collection(ArrayList::new , PacketCodecs.STRING), SceneOptions::slowAnim,
             PacketCodecs.collection(ArrayList::new, PacketCodecs.STRING), SceneOptions::fastAnim,
             PacketCodecs.STRING, SceneOptions::cumAnim,
+            PacketCodecs.BOOLEAN, SceneOptions::needsToStrip,
             PacketCodecs.BOOLEAN, SceneOptions::isBedScene,
             PacketCodecs.FLOAT, SceneOptions::bedOffset,
             PacketCodecs.collection(ArrayList::new, PacketCodecs.STRING), SceneOptions::bedIdle,
-            PacketCodecs.BOOLEAN, SceneOptions::needsToStrip,
-            PacketCodecs.VAR_INT, SceneOptions::requiredRelationshipLevel,
             SceneOptions::new
     );
 }